Added a comment (see description)
This commit's the beginning of a sweep of the code to check for any possible issues of incompatibility between D's GC and Janet's expectations. For the most part, using nogc everywhere should prevent this, but the fact that I missed registerFunc in a previous commit suggests to me that I need to go through and make absolute sure that I'm not handing Janet any memory that might end up being invalid due to the GC at a later time.